:hammer: Refactor task module for multi-platform impl
