+++ composeApp/src/commonMain/kotlin/com/crosspaste/task/CleanPasteTaskExecutor.kt
@@ -9,10 +9,8 @@ import com.crosspaste.dao.task.TaskType
 import com.crosspaste.exception.StandardErrorCode
 import com.crosspaste.net.clientapi.createFailureResult
 import com.crosspaste.task.extra.BaseExtraInfo
-import com.crosspaste.utils.DateUtils
-import com.crosspaste.utils.DesktopTaskUtils
-import com.crosspaste.utils.DesktopTaskUtils.createFailurePasteTaskResult
 import com.crosspaste.utils.getDateUtils
+import com.crosspaste.utils.getTaskUtils
 import io.github.oshai.kotlinlogging.KLogger
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.realm.kotlin.types.RealmInstant
@@ -26,7 +24,9 @@ class CleanPasteTaskExecutor(
 
     private val logger: KLogger = KotlinLogging.logger {}
 
-    private val dateUtils: DateUtils = getDateUtils()
+    private val dateUtils = getDateUtils()
+
+    private val taskUtils = getTaskUtils()
 
     override val taskType: Int = TaskType.CLEAN_PASTE_TASK
 
@@ -45,8 +45,8 @@ class CleanPasteTaskExecutor(
                     pasteDao.markDeleteByCleanTime(fileCleanTimeInstant, PasteType.FILE)
                 }
             } catch (e: Throwable) {
-                val baseExtraInfo = DesktopTaskUtils.getExtraInfo(pasteTask, BaseExtraInfo::class)
-                return createFailurePasteTaskResult(
+                val baseExtraInfo = taskUtils.getExtraInfo(pasteTask, BaseExtraInfo::class)
+                return taskUtils.createFailurePasteTaskResult(
                     logger = logger,
                     retryHandler = { baseExtraInfo.executionHistories.size < 2 },
                     startTime = pasteTask.modifyTime,
@@ -69,8 +69,8 @@ class CleanPasteTaskExecutor(
                     }
                 }
             } catch (e: Throwable) {
-                val baseExtraInfo = DesktopTaskUtils.getExtraInfo(pasteTask, BaseExtraInfo::class)
-                return createFailurePasteTaskResult(
+                val baseExtraInfo = taskUtils.getExtraInfo(pasteTask, BaseExtraInfo::class)
+                return taskUtils.createFailurePasteTaskResult(
                     logger = logger,
                     retryHandler = { baseExtraInfo.executionHistories.size < 2 },
                     startTime = pasteTask.modifyTime,
