+++ composeApp/src/commonMain/kotlin/com/crosspaste/task/DeletePasteTaskExecutor.kt
@@ -6,13 +6,14 @@ import com.crosspaste.dao.task.TaskType
 import com.crosspaste.exception.StandardErrorCode
 import com.crosspaste.net.clientapi.createFailureResult
 import com.crosspaste.task.extra.BaseExtraInfo
-import com.crosspaste.utils.DesktopTaskUtils
-import com.crosspaste.utils.DesktopTaskUtils.createFailurePasteTaskResult
+import com.crosspaste.utils.getTaskUtils
 import io.github.oshai.kotlinlogging.KotlinLogging
 
 class DeletePasteTaskExecutor(private val pasteDao: PasteDao) : SingleTypeTaskExecutor {
 
     private val logger = KotlinLogging.logger {}
+
+    private val taskUtils = getTaskUtils()
 
     override val taskType: Int = TaskType.DELETE_PASTE_TASK
 
@@ -21,12 +22,12 @@ class DeletePasteTaskExecutor(private val pasteDao: PasteDao) : SingleTypeTaskEx
             pasteDao.deletePasteData(pasteTask.pasteDataId!!)
             return SuccessPasteTaskResult()
         } catch (e: Throwable) {
-            return createFailurePasteTaskResult(
+            return taskUtils.createFailurePasteTaskResult(
                 logger = logger,
                 retryHandler = { false },
                 startTime = pasteTask.modifyTime,
                 fails = listOf(createFailureResult(StandardErrorCode.DELETE_TASK_FAIL, e)),
-                extraInfo = DesktopTaskUtils.getExtraInfo(pasteTask, BaseExtraInfo::class),
+                extraInfo = taskUtils.getExtraInfo(pasteTask, BaseExtraInfo::class),
             )
         }
     }
