+++ composeApp/src/commonMain/kotlin/com/crosspaste/task/PullFileTaskExecutor.kt
@@ -20,15 +20,15 @@ import com.crosspaste.presist.FilesIndexBuilder
 import com.crosspaste.sync.SyncManager
 import com.crosspaste.task.extra.PullExtraInfo
 import com.crosspaste.utils.DateUtils
-import com.crosspaste.utils.DesktopTaskUtils
-import com.crosspaste.utils.DesktopTaskUtils.createFailurePasteTaskResult
 import com.crosspaste.utils.FileUtils
 import com.crosspaste.utils.buildUrl
 import com.crosspaste.utils.getDateUtils
 import com.crosspaste.utils.getFileUtils
+import com.crosspaste.utils.getTaskUtils
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.ktor.http.*
 import io.ktor.utils.io.*
+import kotlinx.datetime.Clock
 import org.mongodb.kbson.ObjectId
 
 class PullFileTaskExecutor(
@@ -51,10 +51,12 @@ class PullFileTaskExecutor(
         private val fileUtils: FileUtils = getFileUtils()
     }
 
+    private val taskUtils = getTaskUtils()
+
     override val taskType: Int = TaskType.PULL_FILE_TASK
 
     override suspend fun doExecuteTask(pasteTask: PasteTask): PasteTaskResult {
-        val pullExtraInfo: PullExtraInfo = DesktopTaskUtils.getExtraInfo(pasteTask, PullExtraInfo::class)
+        val pullExtraInfo: PullExtraInfo = taskUtils.getExtraInfo(pasteTask, PullExtraInfo::class)
 
         pasteDao.getPasteData(pasteTask.pasteDataId!!)?.let { pasteData ->
             val fileItems = pasteData.getPasteAppearItems().filter { it is PasteFiles }
@@ -185,7 +187,7 @@ class PullFileTaskExecutor(
         pasteData: PasteData,
         pullExtraInfo: PullExtraInfo,
         fails: Collection<FailureResult>,
-        startTime: Long = System.currentTimeMillis(),
+        startTime: Long = Clock.System.now().toEpochMilliseconds(),
     ): PasteTaskResult {
         val needRetry = pullExtraInfo.executionHistories.size < 3
 
@@ -194,7 +196,7 @@ class PullFileTaskExecutor(
             pasteboardService.clearRemotePasteboard(pasteData)
         }
 
-        return createFailurePasteTaskResult(
+        return taskUtils.createFailurePasteTaskResult(
             logger = logger,
             retryHandler = { needRetry },
             startTime = startTime,
